requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==23.1.0
auto-py-to-exe==2.35.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.5.7
cffi==1.15.1
Eel==0.16.0
exceptiongroup==1.1.1
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
lxml==4.9.2
outcome==1.2.0
pefile==2023.2.7
pycparser==2.21
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PySocks==1.7.1
python-docx==0.8.11
python-dotenv==1.0.0
pywin32-ctypes==0.2.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
whichcraft==0.6.1
wsproto==1.2.0
zope.event==4.6
zope.interface==6.0
